\docType{class}
\name{format.tables-class}
\alias{format.tables}
\alias{format.tables-class}
\title{format.tables class}
\arguments{
  \item{data}{data frame, argument used to initialise
  class}

  \item{styles}{character vector with styles , argument
  used to initialise class}

  \item{column.names}{column names of the data , argument
  used to initialise class}

  \item{names.style}{style for the column names, argument
  used to initialise class}

  \item{header}{list of key, value for header information,
  e.g. caption, subcaption, width etc. , argument used to
  initialise class}

  \item{names.note}{note for the column names, argument
  used to initialise class}

  \item{notes}{notes for the data, argument used to
  initialise class}

  \item{file}{the name of the file which the data are to be
  read from}

  \item{...}{arguments passed to
  \\code{\\link[format.tables:read]{read}}}
}
\description{
Class for format.tables
}

