% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read}
\alias{read}
\title{read data from format.tables format}
\arguments{
\item{file}{the name of the file which the data are to be read from, for more information see \code{\link{read.table}}}

\item{has.column.names}{does the table contain a row with column names that are of a different type than the data?}

\item{convert.data}{should the data be processed with \code{\link{type.convert}}?}

\item{dec}{decimal separator, used for type conversion}

\item{sep}{the field separator character. Values on each line of the file are separated by this character, default is ","}

\item{na.strings}{passed to \code{\link{read.table}}, default "NA" and blanks ("") are always treated as NAs}

\item{...}{parameters passed to read.table}
}
\description{
Read data from format.table format which is a special formatted csv
}

